+++ js/api.js
@@ -10,7 +10,6 @@ const getData = (onSuccess, onFail) => {
     });
 };
 
-
 const sendData = (onSuccess, onFail, body) => {
   fetch(
     'https://26.javascript.pages.academy/keksobooking',
